+++ src/dbtype/mod.rs
@@ -5,6 +5,9 @@
 // types are serialized as JSON.
 
 #[macro_use]
+mod hex16;
+
+#[macro_use]
 mod impl_name_macro;
 
 #[cfg(test)]
@@ -24,6 +27,8 @@ mod document_id;
 mod document_name;
 mod error_response;
 mod revision;
+mod root;
+mod uuid;
 mod view_function;
 mod view_name;
 mod view_result;
@@ -43,6 +48,8 @@ pub use self::document_id::DocumentId;
 pub use self::document_name::DocumentName;
 pub use self::error_response::ErrorResponse;
 pub use self::revision::Revision;
+pub use self::root::{Root, RootBuilder, Vendor};
+pub use self::uuid::Uuid;
 pub use self::view_function::{ViewFunction, ViewFunctionBuilder, ViewFunctionMap};
 pub use self::view_name::ViewName;
 pub use self::view_result::ViewResult;